Ensure each process initializes logger
